refactor: moved manageGitLabTokens command to global
